requirements.txt
rembg==2.0.50
onnxruntime==1.15.1
flask
pillow